+++ back/src/bsvhu/resolvers/BsvhuMetadata.ts
@@ -6,7 +6,9 @@ import {
   BsvhuStatus
 } from "../../generated/graphql/types";
 import { getBsvhuOrNotFound } from "../database";
-import { validateBsvhu } from "../validation";
+import { parseBsvhu } from "../validation";
+import { prismaToZodBsvhu } from "../validation/helpers";
+import { SignatureTypeInput } from "../../generated/graphql/types";
 
 const bsvhuMetadataResolvers: BsvhuMetadataResolvers = {
   errors: async (
@@ -17,30 +19,15 @@ const bsvhuMetadataResolvers: BsvhuMetadataResolvers = {
     const validationMatrix = [
       {
         skip: metadata.status !== "INITIAL",
-        requiredFor: "EMISSION",
-        context: {
-          emissionSignature: true,
-          transportSignature: false,
-          operationSignature: false
-        }
+        requiredFor: "EMISSION"
       },
       {
         skip: ["PROCESSED", "REFUSED", "SENT"].includes(metadata.status),
-        requiredFor: "TRANSPORT",
-        context: {
-          emissionSignature: true,
-          transportSignature: true,
-          operationSignature: false
-        }
+        requiredFor: "TRANSPORT"
       },
       {
         skip: ["PROCESSED", "REFUSED"].includes(metadata.status),
-        requiredFor: "OPERATION",
-        context: {
-          emissionSignature: true,
-          transportSignature: true,
-          operationSignature: true
-        }
+        requiredFor: "OPERATION"
       }
     ];
 
@@ -50,16 +37,16 @@ const bsvhuMetadataResolvers: BsvhuMetadataResolvers = {
 
     // import transporterReceipt that will be completed after transporter signature
     const transporterReceipt = await getTransporterReceipt(prismaBsvhu);
+    const zodBsvhu = prismaToZodBsvhu({
+      ...prismaBsvhu,
+      ...transporterReceipt
+    });
 
-    for (const { context, requiredFor } of filteredValidationMatrix) {
+    for (const { requiredFor } of filteredValidationMatrix) {
       try {
-        await validateBsvhu(
-          {
-            ...prismaBsvhu,
-            ...transporterReceipt
-          },
-          context
-        );
+        parseBsvhu(zodBsvhu, {
+          currentSignatureType: requiredFor as SignatureTypeInput
+        });
         return [];
       } catch (errors) {
         return errors.inner?.map((e: ValidationError) => {