+++ back/src/common/validation/zod/refinement.ts
@@ -1,4 +1,4 @@
-import { z } from "zod";
+import { RefinementCtx, z } from "zod";
 import {
   isCollector,
   isTransporter,
@@ -8,6 +8,8 @@ import {
 } from "../../../companies/validation";
 import { prisma } from "@td/prisma";
 import { Company, CompanyVerificationStatus } from "@prisma/client";
+import { getOperationModesFromOperationCode } from "../../operationModes";
+import { CompanyRole } from "./schema";
 
 const { VERIFY_COMPANY } = process.env;
 
@@ -19,11 +21,15 @@ export async function isTransporterRefinement(
     siret: string | null | undefined;
     transporterRecepisseIsExempted: boolean;
   },
-  ctx
+  ctx: RefinementCtx
 ) {
   if (transporterRecepisseIsExempted) return;
 
-  const company = await refineSiretAndGetCompany(siret, ctx);
+  const company = await refineSiretAndGetCompany(
+    siret,
+    ctx,
+    CompanyRole.Transporter
+  );
 
   if (company && !isTransporter(company)) {
     ctx.addIssue({
@@ -38,7 +44,8 @@ export async function isTransporterRefinement(
 }
 export async function refineSiretAndGetCompany(
   siret: string | null | undefined,
-  ctx
+  ctx: RefinementCtx,
+  companyRole?: CompanyRole
 ): Promise<Company | null> {
   if (!siret) return null;
   const company = await prisma.company.findUnique({
@@ -48,13 +55,18 @@ export async function refineSiretAndGetCompany(
   if (company === null) {
     ctx.addIssue({
       code: z.ZodIssueCode.custom,
-      message: `L'établissement avec le SIRET ${siret} n'est pas inscrit sur Trackdéchets`
+      message: `${
+        companyRole ? `${companyRole} : ` : ""
+      }L'établissement avec le SIRET ${siret} n'est pas inscrit sur Trackdéchets`
     });
   }
 
   return company;
 }
-export const isRegisteredVatNumberRefinement = async (vatNumber, ctx) => {
+export const isRegisteredVatNumberRefinement = async (
+  vatNumber: string | null | undefined,
+  ctx: RefinementCtx
+) => {
   if (!vatNumber) return;
   const company = await prisma.company.findUnique({
     where: { vatNumber }
@@ -78,12 +90,26 @@ export const isRegisteredVatNumberRefinement = async (vatNumber, ctx) => {
 };
 export async function isDestinationRefinement(
   siret: string | null | undefined,
-  ctx,
+  ctx: RefinementCtx,
+  role: "DESTINATION" | "WASTE_VEHICLES" = "DESTINATION",
   isExemptedFromVerification?: (destination: Company | null) => boolean
 ) {
-  const company = await refineSiretAndGetCompany(siret, ctx);
+  const company = await refineSiretAndGetCompany(
+    siret,
+    ctx,
+    CompanyRole.Destination
+  );
 
-  if (
+  if (company && role === "WASTE_VEHICLES" && !isWasteVehicles(company)) {
+    return ctx.addIssue({
+      code: z.ZodIssueCode.custom,
+      message:
+        `L'installation de destination avec le SIRET "${siret}" n'est pas inscrite` +
+        ` sur Trackdéchets en tant qu'installation de traitement de VHU. Cette installation ne peut` +
+        ` donc pas être visée sur le bordereau. Veuillez vous rapprocher de l'administrateur de cette installation pour qu'il` +
+        ` modifie le profil de l'établissement depuis l'interface Trackdéchets Mon Compte > Établissements`
+    });
+  } else if (
     company &&
     !isCollector(company) &&
     !isWasteProcessor(company) &&
@@ -115,5 +141,33 @@ export async function isDestinationRefinement(
         `Le compte de l'installation de destination ou d’entreposage ou de reconditionnement prévue` +
         ` avec le SIRET ${siret} n'a pas encore été vérifié. Cette installation ne peut pas être visée sur le bordereau.`
     });
+  }
+}
+
+export function destinationOperationModeRefinement(
+  destinationOperationCode: string | null | undefined,
+  destinationOperationMode: string | null | undefined,
+  ctx: RefinementCtx
+) {
+  if (destinationOperationCode) {
+    const modes = getOperationModesFromOperationCode(destinationOperationCode);
+
+    if (modes.length && !destinationOperationMode) {
+      return ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: "Vous devez préciser un mode de traitement"
+      });
+    } else if (
+      (modes.length &&
+        destinationOperationMode &&
+        !modes.includes(destinationOperationMode)) ||
+      (!modes.length && destinationOperationMode)
+    ) {
+      return ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message:
+          "Le mode de traitement n'est pas compatible avec l'opération de traitement choisie"
+      });
+    }
   }
 }